+++ src/webgl/shaderity_shaders/DetectHighLuminanceAndCorrectShader/DetectHighLuminanceAndCorrectShader.frag
@@ -18,7 +18,7 @@ void main ()
 
   float luminance = dot(baseColor.rgb, vec3(0.2126, 0.7152, 0.0722));
 
-  float luminanceCriterion = get_luminanceCriterion(uint(materialSID), 0u);
+  float luminanceCriterion = get_luminanceCriterion(materialSID, 0u);
 
   baseColor.rgb = mix(vec3(0.0), baseColor.rgb, (luminance - luminanceCriterion) / luminanceCriterion);
   baseColor.a = 1.0;
