+++ src/webgl/shaderity_shaders/VarianceShadowMapDecodeClassicShader/VarianceShadowMapDecodeClassicShader.vert
@@ -22,7 +22,7 @@ void main(){
 
   mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
   mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
-  mat4 viewMatrix = get_viewMatrix(uint(cameraSID));
+  mat4 viewMatrix = get_viewMatrix(cameraSID);
 
   bool isSkinning = false;
   isSkinning = processGeometry(
@@ -39,7 +39,7 @@ void main(){
     v_normal_inWorld
   );
 
-  mat4 projectionMatrix = get_projectionMatrix(uint(cameraSID));
+  mat4 projectionMatrix = get_projectionMatrix(cameraSID);
   gl_Position = projectionMatrix * viewMatrix * v_position_inWorld;
 
 
@@ -49,7 +49,7 @@ void main(){
   v_texcoord_0 = a_texcoord_0;
 
   // Shadow mapping
-  mat4 lightViewProjectionMatrix = get_lightViewProjectionMatrix(uint(materialSID), 0u);
+  mat4 lightViewProjectionMatrix = get_lightViewProjectionMatrix(materialSID, 0u);
   v_projPosition_from_light = lightViewProjectionMatrix * v_position_inWorld;
 
   // Following tMatrix is based on https://wgld.org/d/webgl/w051.html
