+++ src/webgl/shaderity_shaders/common/outputSrgb.glsl
@@ -1,2 +1,2 @@
-float makeOutputSrgb = float(get_makeOutputSrgb(uint(materialSID), 0u));
+float makeOutputSrgb = float(get_makeOutputSrgb(materialSID, 0u));
 rt0.rgb = mix(rt0.rgb, linearToSrgb(rt0.rgb), makeOutputSrgb);
